[GEOT-7593] Updating to PMD 7.13

+++ modules/library/metadata/src/main/java/org/geotools/util/Base64.java
@@ -841,7 +841,7 @@ public class Base64 {
         // GZIP Magic Two-Byte Number: 0x8b1f (35615)
         if (bytes != null && bytes.length >= 4) {
 
-            int head = ((int) bytes[0] & 0xff) | ((bytes[1] << 8) & 0xff00);
+            int head = (bytes[0] & 0xff) | ((bytes[1] << 8) & 0xff00);
             if (java.util.zip.GZIPInputStream.GZIP_MAGIC == head) {
                 byte[] buffer = new byte[2048];
                 int length = 0;
